+++ ElvUI_MerathilisUI/Core/Changelog/6.22.lua
@@ -1,8 +1,13 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[622] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "27.07.2024",
+	FIXES = {
+		"[Announcement]: Added back some Options, whoooopppsiii",
+		"[Objective Tracker]: A lot of work on the Objective Tracker. Thx fang2hou and Blinkii",
+	},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[Widget Skin]: Some gradient look on the Widget Skin",
+	},
 }