+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### Changes ###
 
-  * [Objective Tracker]: Fix copy & paste whoooopsiii
+  * [Announcement]: Added back some Options, whoooopppsiii
+  * [Objective Tracker]: A lot of work on the Objective Tracker. Thx fang2hou and Blinkii
+  * [Widget Skin]: Some gradient look on the Widget Skin